requirements.txt
Flask>=3.0.0
flask-cors
google-cloud-vision
requests
